move main components location
