+++ src/pages/main/components/join/join-button.tsx
