+++ src/pages/main/components/join/join-intro.tsx
