+++ src/pages/main/components/welcome/components/welcome-title.tsx
